# swagger/api/models/productShema.js
const mongoose = require('mongoose');

const ProductSchema = new mongoose.Schema({
    prod_name: String,
    prod_brand: String,
    prod_company: {type: String, required: true},
    prod_price: String,
    prod_isbn: String,
    lastModifiedDate: { type: String, default: Date.now },
    reviews: {
        id: Number,
        user_id: Number,
        text: String
    }
});

const Product = mongoose.model('Product', ProductSchema);

module.exports = Product;